--- a/test_data/irobot_hand/model.config
+++ b/test_data/irobot_hand/model.config
@@ -1,0 +1,16 @@
+<?xml version='1.0'?>
+<model>
+  <name>iRobot Hand</name>
+  <version>0.1.0</version>
+  <sdf version='1.3'>model-1_3.sdf</sdf>
+  <sdf version='1.4'>model-1_4.sdf</sdf>
+  <sdf version='1.5'>model.sdf</sdf>
+
+  <author>
+    <name>OSRF</name>
+    <email>hsu@osrfoundation.org</email>
+  </author>
+  <description>
+    iRobot hand model.
+  </description>
+</model>