--- a/test_data/pioneer2dx/model.config
+++ b/test_data/pioneer2dx/model.config
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+
+<model>
+  <name>Pioneer 2DX</name>
+  <version>1.0</version>
+  <sdf version="1.2">model-1_2.sdf</sdf>
+  <sdf version="1.3">model-1_3.sdf</sdf>
+  <sdf version="1.4">model-1_4.sdf</sdf>
+  <sdf version="1.5">model.sdf</sdf>
+
+  <author>
+    <name>Nate Koenig</name>
+    <email>nate@osrfoundation.org</email>
+  </author>
+
+  <description>
+    A model of the pioneer 2 dx robot.
+  </description>
+
+  <depend>
+  </depend>
+</model>