--- a/test_data/simple_arm/model.config
+++ b/test_data/simple_arm/model.config
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+
+<model>
+  <name>Simple Arm</name>
+  <version>1.0</version>
+  <sdf version="1.2">model-1_2.sdf</sdf>
+  <sdf version="1.3">model-1_3.sdf</sdf>
+  <sdf version="1.4">model-1_4.sdf</sdf>
+  <sdf version="1.5">model.sdf</sdf>
+
+  <author>
+    <name>John Hsu</name>
+    <email>hsu@osrfoundation.org</email>
+  </author>
+
+  <description>
+    A simple SCARA arm.
+  </description>
+</model>