Depends on:
- [x] https://github.com/rock-gui/gui-control_ui/pull/8

In addition to the SDF support _per se_, there's a change in behavior stuck in there. Instead of running the SmoothVisitor, which can take a lot of time on load-time, the plugin only runs the VBOVisitor, which ensures that OSG uses VBO on modern OpenGL (a.k.a. "everywhere" these days). The reason is that:

- the smooth visitor ended up very expensive in our scenes. This led to serious loading times (in the 30s to 1 min) range. The workflow after this commit was to preprocess mesh files into a .osgb file in the build system (in our case, the Rakefile from our bundle), and have the plugin auto-load a `<filename>.osgb` instead of `<filename>` when requested.
- the VBO visitor divided the rendering times by an order of magnitude (and allowed the usage of opengl debuggers, because a 1 min rendering was logged as less than a GB instead of 15-20GB)

Moreover, this significantly expands the functionality of rock-roboviz to allow spawning a control UI (from gui/control_ui).