@planthaber 2017-08-22
rock-roboviz:99: syntax error, unexpected keyword_elsif, expecting keyword_end
elsif use_test_gui