@doudou 2017-08-22
Fixed. The fix depends on changes in https://github.com/rock-gui/gui-control_ui/pull/8 as well