@maltewi 2017-09-18
I've tested the PR. While welcoming it in principle I see two things. The first thing seems to be a bug, the second I don't know, but maybe you?

1) The mechanism of automatically loading .osgb files seems not to work.
I've tested it with the spaceclimber_multi_visuals data set. Therefor I've placed `.osgb` version of the mesh files intot the same folder as the original `.obj` files. Calling `rock-roboviz SpaceClimber.urdf` within the est_data/spaceclimber_multi_visuals` folder still results in the .obj files being loaded.
Can be seen by loading take long time as compared to loading a .urdf with Search and replace '.obj' --> '.osgb'.

   a) I'm not a real friend of this feature. It creates some unclarity in behaviour. I see people getting confused over the fact, that a different file is loaded than the one which is defined in the URDF/SDF model file. An alternative is proposed in the discussion of #14. Please have a look.

2) I have significantly worse rendering performance on my PC. The test, again the spaceclimber_multi_visuals data set. 
My test:
I performed the optimization offline using the program `optimize_meshes` which I pushed to the PR #14: 
```
SMOOTH=true optimize_meshes *.obj
```
Now loaded a version of the URDF file, that loads the pre-processed `.osgb` files. This is the result:

|                | original file (.obj)  |  .osgb-version after `optimize_meshes` |
| ------------ | ---------- | ------------- |
| this pr      | ~8.5 fps | ~11 fps |
| old master | ~25 fps | ~29 fps |

Graphics board: Intel® HD Graphics 520 (Skylake GT2)
Do you have an idea what's the reason?