@doudou 2017-09-18
> The mechanism of automatically loading .osgb files seems not to work.

My guess is that we've not implemented it for URDF.

> a) I'm not a real friend of this feature. It creates some unclarity in behaviour

I agree that it's not perfect. The problem is that having a "base" model and then having to post-process (what you propose) also adds significant complexity in understanding. I personally found the convention of looking for an osgb file along the main mesh file a better trade-off.

On the SDF side, Gazebo does not support osgb, so we get the problem that we can't have the osgb file in the SDF file directly.