@doudou 2017-09-18
Just did some tests. Reading the `optimize_meshes` script, it seems that I was not clear enough in my explanations. The convention is to add the extension, i.e. visual_leg0_motor1_an.obj becomes visual_leg0_motor1_an.obj.osgb.

With the osgb files generated with the expected name, I get the following performance for the spaceclimber_multi_visuals on my i7-4700MQ / Intel HD 4600

|               | .obj   | .osgb |
|------------|---------|---------|
| This PR | 2 fps | 29 fps |
| master   | 20 fps | N/A |

The drop in performance with the .obj is expected if the .obj itself is not optimized. I originally did not elect to still run the optimizer on the non-osgb file as it was causing us huge loading times. But then, we now get slow renders. Don't know what would be the best convention here. I don't mind so much either way, as I'm definitely going to create a "asset pipeline" documentation of sorts that allows for the creation of the osgb files.

In addition, loading the file is also significantly faster (from script startup to first visual, there's ~5s of initializing the qt plugins) 

|               | .obj   | .osgb |
|------------|---------|---------|
| This PR | 40s | 7s |
| master   | 24s | N/A |

(FYI, we've originally done this work because loading a ground mesh was taking minutes).