@doudou 2017-09-18
Actually, the fps is not representative. Vizkit3d renders at 30fps no matter what, and it does not account for separate CPU and GPU usage. Below are rendering times:

| Test | CPU time | GPU time |
|-------|--------------|--------------|
| .obj, master | 32ms | 32ms |
| .obj, PR | 450ms | 450ms |
| .osgb, PR | 4ms | 12ms |

Note that the .obj version could definitely be as fast as the .osgb version. The main difference is the usage of vertex buffers, and what optimizers are run. What the `.osgb` really adds is better load times (regardless of how much optimization needs to be done on the mesh).