@maltewi 2017-09-25
Yes, loading time is faster with osgb. I personally use the format for this reason as well. Removing the optimization step at load time improves loading even more. I think, for that reason it's a valid idea to remove the online optimization, if a tool for performing the optimization is provided at the same time (which is done with `optimize_meshes`).

My problem with this PR is that I can't get same rendering performance as with old master (2. in my previous answer). Maybe this is related to the offline optimization (then same performance should be possible by adjusting parameters in the `optimize_meshes` script). But maybe it's related to something else in the PR. That we've to find out what's causing the slow down.

Here's how I experience the slow down: I apply optimization offline with the `optimize_meshes` script and load a URDF file that loads the osgb versions of the meshes. After loading I press 's' until I see the debug info graph. It shows:
   * ~11fps
   * ~68ms Draw
   * ~63ms GPU 
I also see and feel the slow performance when adjusting the camera.

Comparing with old master and the .obj files (optimization applied, when loading the files) gives:
   * ~25 fps
   * ~17ms Draw
   * ~15ms GPU

As you can see from the numbers, that's a significant loss in performance, which I think is not okay.