@doudou 2017-09-25
> As you can see from the numbers, that's a significant loss in performance, which I think is not okay.

I completely understand and agree. As you can see from my numbers, I don't have the same numbers at all. So, next question, what's the difference ?

Just to be sure that we're talking about the same thing:
 - the non-osgb case performance *is* dropping, which is something I would expect. It might be something we want to avoid as well, given that after this PR everyone will get the performance drop.
 - the osgb performance is not present in the tests I did. The `optimize_meshes` script is buggy, and does not generate the expected file names (which causes the .osgb to not be loaded). See https://github.com/rock-gui/gui-robot_model/pull/15 for a fix.